requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=3.0.0
transformers>=4.40.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pymongo>=4.6.0
APScheduler>=3.10.0
pydantic-ai[groq,google,huggingface]>=0.1.0
openai>=1.0.0
nest_asyncio>=1.6.0
